Add spreadsheet diaries to AI Assistant
